Support running the dataset loader outside the ansible control host
I've been chasing intermittent issues with regard to dropped packets or connections breaking python3 or urllib3 whereas it's not been an issue with python2.7 and urllib.  This has manifested itself as "incomplete read" exceptions being thrown inside the URI module.  This adds support so that when using the import_role module for the dataset loader it can delegate the default host upon which it runs.  This means that the URI calls instead of referencing a normal TO url, would reference localhost instead.  The dataset loader does however have a few dependencies on the ansible control host for SSL certificates on delivery services, so we have to add an explicit delegate_to whereas normally it would be implicit.