Switch initializing SOLR to curl
Riak has a known problems with TLS1.2.  By switching these interactions
to curl explicitly we can disable it at the client since this isn't
normally exposed to ansible.  It may or may not work as it was depending
on exactly what combinations of python, urllib, openssl, and ansible you
were using.  This should generally be more reliable though.